Refactored h3.py into 3 scripts: constants, components not specific to h3, and components specific to h3
