+++ vector2dggs/h3.py
@@ -1,81 +1,28 @@
-import errno
-import logging
-import os
 import multiprocessing
-from multiprocessing.dummy import Pool
-from pathlib import Path, PurePath
-import shutil
 import sys
 import tempfile
-from typing import Union
-from urllib.parse import urlparse
-import warnings
-
 import click
 import click_log
-import dask.dataframe as dd
-import dask_geopandas as dgpd
-import geopandas as gpd
-import h3pandas
-import pandas as pd
 import pyproj
-from shapely.geometry import GeometryCollection
-import sqlalchemy
-from tqdm import tqdm
-from tqdm.dask import TqdmCallback
 
-from . import katana
+import pandas as pd
+import geopandas as gpd
+import h3pandas  # Necessary import despite lack of explicit use
+
+from typing import Union
+from pathlib import Path
+
+import vector2dggs.constants as const
+import vector2dggs.common as common
+
 from vector2dggs import __version__
 
-LOGGER = logging.getLogger(__name__)
-click_log.basic_config(LOGGER)
-click_log.ColorFormatter.colors['info'] = dict(fg="green")
-MIN_H3, MAX_H3 = 0, 15
 
-warnings.filterwarnings(
-    "ignore"
-)  # This is to filter out the polyfill warnings when rows failed to get indexed at a resolution, can be commented out to find missing rows
+def h3_secondary_index(df: gpd.GeoDataFrame, parent_res: int) -> gpd.GeoDataFrame:
+    return df.h3.h3_to_parent(parent_res)
 
 
-DEFAULT_PARENT_OFFSET = 6
-DEFAULT_CHUNK_SIZE = 50
-
-
-class ParentResolutionException(Exception):
-    pass
-
-
-def _get_parent_res(parent_res: Union[None, int], resolution: int):
-    """
-    Uses a parent resolution,
-    OR,
-    Given a target resolution, returns our recommended parent resolution.
-
-    Used for intermediate re-partioning.
-    """
-    return (
-        int(parent_res)
-        if parent_res is not None
-        else max(MIN_H3, (resolution - DEFAULT_PARENT_OFFSET))
-    )
-
-
-def polyfill(
-    pq_in: Path,
-    spatial_sort_col: str,
-    resolution: int,
-    parent_res: Union[None, int],
-    output_directory: str,
-) -> None:
-    """
-    Reads a geoparquet, performs H3 polyfilling (for Polygon),
-    linetracing (for LineString), and writes out to parquet.
-    """
-    df = gpd.read_parquet(pq_in).reset_index().drop(columns=[spatial_sort_col])
-    if len(df.index) == 0:
-        # Input is empty, nothing to polyfill
-        return None
-
+def h3polyfill(df: gpd.GeoDataFrame, resolution: int):
     df_polygon = df[df.geom_type == "Polygon"]
     if len(df_polygon.index) > 0:
         df_polygon = df_polygon.h3.polyfill_resample(
@@ -91,207 +38,23 @@ def polyfill(
         )
         df_linestring = df_linestring[~df_linestring.index.duplicated(keep="first")]
 
-    df = pd.concat(
+    return pd.concat(
         map(
             lambda _df: pd.DataFrame(_df.drop(columns=[_df.geometry.name])),
             [df_polygon, df_linestring],
         )
     )
 
-    if len(df.index) == 0:
-        # Polyfill resulted in empty output (e.g. large cell, small feature)
-        return None
-
-    df.index.rename(f"h3_{resolution:02}", inplace=True)
-    parent_res: int = _get_parent_res(parent_res, resolution)
-    # Secondary (parent) H3 index, used later for partitioning
-    df.h3.h3_to_parent(parent_res).to_parquet(
-        PurePath(output_directory, pq_in.name), engine="auto", compression="ZSTD"
-    )
-    return None
-
-
-def polyfill_star(args) -> None:
-    return polyfill(*args)
-
-
-def _parent_partitioning(
-    input_dir: Path,
-    output_dir: Path,
-    resolution: int,
-    parent_res: Union[None, int],
-    **kwargs,
-) -> None:
-    parent_res: int = _get_parent_res(parent_res, resolution)
-    partition_col = f"h3_{parent_res:02}"
-
-    with TqdmCallback(desc="Repartitioning"):
-        dd.read_parquet(input_dir, engine="pyarrow").to_parquet(
-            output_dir,
-            overwrite=kwargs.get("overwrite", False),
-            engine=kwargs.get("engine", "pyarrow"),
-            partition_on=partition_col,
-            compression=kwargs.get("compression", "ZSTD"),
-        )
-    LOGGER.debug("Parent cell repartitioning complete")
-
-    # Rename output to just be the partition key, suffix .parquet
-    for f in os.listdir(output_dir):
-        os.rename(
-            os.path.join(output_dir, f),
-            os.path.join(output_dir, f.replace(f"{partition_col}=", "") + ".parquet"),
-        )
-
-    return
-
-
-def drop_condition(
-    df: pd.DataFrame,
-    drop_index: pd.Index,
-    log_statement: str,
-    warning_threshold: float = 0.01,
-):
-    LOGGER.debug(log_statement)
-    _before = len(df)
-    df = df.drop(drop_index)
-    _after = len(df)
-    _diff = _before - _after
-    if _diff:
-        log_method = (
-            LOGGER.info if (_diff / float(_before)) < warning_threshold else LOGGER.warn
-        )
-        log_method(f"Dropped {_diff} rows ({_diff/float(_before)*100:.2f}%)")
-    return df
-
-
-def _index(
-    input_file: Union[Path, str],
-    output_directory: Union[Path, str],
-    resolution: int,
-    parent_res: Union[None, int],
-    keep_attributes: bool,
-    chunksize: int,
-    spatial_sorting: str,
-    cut_threshold: int,
-    processes: int,
-    id_field: str = None,
-    cut_crs: pyproj.CRS = None,
-    con: Union[sqlalchemy.engine.Connection, sqlalchemy.engine.Engine] = None,
-    table: str = None,
-    geom_col: str = "geom",
-    overwrite: bool = False,
-) -> Path:
-    """
-    Performs multi-threaded H3 polyfilling on (multi)polygons.
-    """
-
-    if table and con:
-        # Database connection
-        if keep_attributes:
-            q = sqlalchemy.text(f"SELECT * FROM {table}")
-        elif id_field and not keep_attributes:
-            q = sqlalchemy.text(f"SELECT {id_field}, {geom_col} FROM {table}")
-        else:
-            q = sqlalchemy.text(f"SELECT {geom_col} FROM {table}")
-        df = gpd.read_postgis(q, con.connect(), geom_col=geom_col).rename_geometry(
-            "geometry"
-        )
-    else:
-        # Read file
-        df = gpd.read_file(input_file)
-
-    if cut_crs:
-        df = df.to_crs(cut_crs)
-    LOGGER.debug("Cutting with CRS: %s", df.crs)
-
-    if id_field:
-        df = df.set_index(id_field)
-    else:
-        df = df.reset_index()
-        df = df.rename(columns={"index": "fid"}).set_index("fid")
-
-    if not keep_attributes:
-        # Remove all attributes except the geometry
-        df = df.loc[:, ["geometry"]]
-
-    LOGGER.debug("Cutting large geometries")
-    with tqdm(total=df.shape[0], desc='Splitting') as pbar:
-        for index, row in df.iterrows():
-            df.loc[index, "geometry"] = GeometryCollection(
-                katana.katana(row.geometry, cut_threshold)
-            )
-            pbar.update(1)
-
-    LOGGER.debug("Exploding geometry collections and multipolygons")
-    df = (
-        df.to_crs(4326)
-        .explode(index_parts=False)  # Explode from GeometryCollection
-        .explode(index_parts=False)  # Explode multipolygons to polygons
-    ).reset_index()
-
-    drop_conditions = [
-        {
-            "index": lambda frame: frame[
-                (frame.geometry.is_empty | frame.geometry.isna())
-            ],
-            "message": "Considering empty or null geometries",
-        },
-        {
-            "index": lambda frame: frame[
-                (frame.geometry.geom_type != "Polygon")
-                & (frame.geometry.geom_type != "LineString")
-            ],  # NB currently points and other types are lost; in principle, these could be indexed
-            "message": "Considering unsupported geometries",
-        },
-    ]
-    for condition in drop_conditions:
-        df = drop_condition(df, condition["index"](df).index, condition["message"])
-
-    ddf = dgpd.from_geopandas(df, chunksize=max(1, chunksize), sort=True)
-
-    LOGGER.debug("Spatially sorting and partitioning (%s)", spatial_sorting)
-    ddf = ddf.spatial_shuffle(by=spatial_sorting)
-    spatial_sort_col = (
-        spatial_sorting
-        if spatial_sorting == "geohash"
-        else f"{spatial_sorting}_distance"
-    )
-
-    with tempfile.TemporaryDirectory(suffix=".parquet") as tmpdir:
-        with TqdmCallback(desc=f'Spatially partitioning'):
-            ddf.to_parquet(tmpdir, overwrite=True)
-
-        filepaths = list(map(lambda f: f.absolute(), Path(tmpdir).glob("*")))
-
-        # Multithreaded polyfilling
-        LOGGER.debug(
-            "H3 Indexing on spatial partitions by polyfill with H3 resolution: %d",
-            resolution,
-        )
-        with tempfile.TemporaryDirectory(suffix=".parquet") as tmpdir2:
-            with Pool(processes=processes) as pool:
-                args = [
-                    (filepath, spatial_sort_col, resolution, parent_res, tmpdir2)
-                    for filepath in filepaths
-                ]
-                list(tqdm(pool.imap(polyfill_star, args), total=len(args), desc='DGGS indexing'))
-
-            _parent_partitioning(
-                tmpdir2, output_directory, resolution, parent_res, overwrite=overwrite
-            )
-
-    return output_directory
-
 
 @click.command(context_settings={"show_default": True})
-@click_log.simple_verbosity_option(LOGGER)
+@click_log.simple_verbosity_option(common.LOGGER)
 @click.argument("vector_input", required=True, type=click.Path(), nargs=1)
 @click.argument("output_directory", required=True, type=click.Path(), nargs=1)
 @click.option(
     "-r",
     "--resolution",
     required=True,
-    type=click.Choice(list(map(str, range(MIN_H3, MAX_H3 + 1)))),
+    type=click.Choice(list(map(str, range(const.MIN_H3, const.MAX_H3 + 1)))),
     help="H3 resolution to index",
     nargs=1,
 )
@@ -299,14 +62,14 @@ def _index(
     "-pr",
     "--parent_res",
     required=False,
-    type=click.Choice(list(map(str, range(MIN_H3, MAX_H3 + 1)))),
+    type=click.Choice(list(map(str, range(const.MIN_H3, const.MAX_H3 + 1)))),
     help="H3 Parent resolution for the output partition. Defaults to resolution - 6",
 )
 @click.option(
     "-id",
     "--id_field",
     required=False,
-    default=None,
+    default=const.DEFAULTS["id"],
     type=str,
     help="Field to use as an ID; defaults to a constructed single 0...n index on the original feature order.",
     nargs=1,
@@ -316,7 +79,7 @@ def _index(
     "--keep_attributes",
     is_flag=True,
     show_default=True,
-    default=False,
+    default=const.DEFAULTS["k"],
     help="Retain attributes in output. The default is to create an output that only includes H3 cell ID and the ID given by the -id field (or the default index ID).",
 )
 @click.option(
@@ -324,7 +87,7 @@ def _index(
     "--chunksize",
     required=True,
     type=int,
-    default=DEFAULT_CHUNK_SIZE,
+    default=const.DEFAULTS["ch"],
     help="The number of rows per index partition to use when spatially partioning. Adjusting this number will trade off memory use and time.",
     nargs=1,
 )
@@ -332,14 +95,14 @@ def _index(
     "-s",
     "--spatial_sorting",
     type=click.Choice(["hilbert", "morton", "geohash"]),
-    default="hilbert",
+    default=const.DEFAULTS["s"],
     help="Spatial sorting method when perfoming spatial partitioning.",
 )
 @click.option(
     "-crs",
     "--cut_crs",
     required=False,
-    default=None,
+    default=const.DEFAULTS["crs"],
     type=int,
     help="Set the coordinate reference system (CRS) used for cutting large geometries (see `--cur-threshold`). Defaults to the same CRS as the input. Should be a valid EPSG code.",
     nargs=1,
@@ -348,7 +111,7 @@ def _index(
     "-c",
     "--cut_threshold",
     required=True,
-    default=5000,
+    default=const.DEFAULTS["c"],
     type=int,
     help="Cutting up large geometries into smaller geometries based on a target length. Units are assumed to match the input CRS units unless the `--cut_crs` is also given, in which case units match the units of the supplied CRS.",
     nargs=1,
@@ -357,7 +120,7 @@ def _index(
     "-t",
     "--threads",
     required=False,
-    default=7,
+    default=const.DEFAULTS["t"],
     type=int,
     help="Amount of threads used for operation",
     nargs=1,
@@ -366,7 +129,7 @@ def _index(
     "-tbl",
     "--table",
     required=False,
-    default=None,
+    default=const.DEFAULTS["tbl"],
     type=str,
     help="Name of the table to read when using a spatial database connection as input",
     nargs=1,
@@ -375,14 +138,14 @@ def _index(
     "-g",
     "--geom_col",
     required=False,
-    default="geom",
+    default=const.DEFAULTS["g"],
     type=str,
     help="Column name to use when using a spatial database connection as input",
     nargs=1,
 )
 @click.option(
     "--tempdir",
-    default=tempfile.tempdir,
+    default=const.DEFAULTS["tempdir"],
     type=click.Path(),
     help="Temporary data is created during the execution of this program. This parameter allows you to control where this data will be written.",
 )
@@ -411,46 +174,22 @@ def h3(
     VECTOR_INPUT is the path to input vector geospatial data.
     OUTPUT_DIRECTORY should be a directory, not a file or database table, as it will instead be the write location for an Apache Parquet data store.
     """
-    tempfile.tempdir = tempdir
-    if parent_res is not None and not int(parent_res) < int(resolution):
-        raise ParentResolutionException(
-            "Parent resolution ({pr}) must be less than target resolution ({r})".format(
-                pr=parent_res, r=resolution
-            )
-        )
-    con: sqlalchemy.engine.Connection = None
-    scheme: str = urlparse(vector_input).scheme
-    if bool(scheme) and scheme != "file":
-        # Assume database connection
-        con = sqlalchemy.create_engine(vector_input)
-    elif not Path(vector_input).exists():
-        if not scheme:
-            LOGGER.error(
-                f"Input vector {vector_input} does not exist, and is not recognised as a remote URI"
-            )
-            raise FileNotFoundError(
-                errno.ENOENT, os.strerror(errno.ENOENT), vector_input
-            )
-        vector_input = str(vector_input)
-    else:
-        vector_input = Path(vector_input)
+    tempfile.tempdir = tempdir if tempdir is not None else tempfile.tempdir
 
-    output_directory = Path(output_directory)
-    outputexists = os.path.exists(output_directory)
-    if outputexists and not overwrite:
-        raise FileExistsError(
-            f"{output_directory} already exists; if you want to overwrite this, use the -o/--overwrite flag"
-        )
-    elif outputexists and overwrite:
-        LOGGER.warn(f"Overwriting the contents of {output_directory}")
-        shutil.rmtree(output_directory)
-    output_directory.mkdir(parents=True, exist_ok=True)
+    common.check_resolutions(resolution, parent_res)
+
+    con, vector_input = common.db_conn_and_input_path(vector_input)
+    output_directory = common.resolve_output_path(output_directory, overwrite)
 
     if cut_crs is not None:
         cut_crs = pyproj.CRS.from_user_input(cut_crs)
 
     try:
-        _index(
+        # TODO: needs to be common._index(...)
+        common._index(
+            "h3",
+            h3polyfill,
+            h3_secondary_index,
             vector_input,
             output_directory,
             int(resolution),
